requirements.txt
python-dotenv
jira
